LayoutTests/platform/mac/accessibility/aria-liveregions-notifications.html

<!DOCTYPE HTML PUBLIC "-//IETF//DTD HTML//EN">
<html>
<head>
<script src="../../../resources/js-test-pre.js"></script>
</head>
<body id="body">

<div role="group" tabindex=0 id="liveregion" aria-live="polite" aria-relevant="additions">
<h3 id="innerlive">text</h3>
<img src="test.gif" width=100 height=100 alt="alt text" tabindex=0 id="image">
</div>

<p id="description"></p>
<div id="console"></div>

<script>
    var jsTestIsAsync = true;

    description("This tests that ARIA live regions are sending out the correct notifications. We perform four operations (add, remove, change text, change alt tag), each one should trigger a live region notification");

    var liveRegion = 0;
    var liveRegionChangeCount = 0;
    function ariaCallback(notification) {
        if (notification == "AXLiveRegionChanged") {
            liveRegionChangeCount++;
        }
    }

    if (window.accessibilityController) {

        liveRegion = accessibilityController.accessibleElementById("liveregion");

        var addedNotification = liveRegion.addNotificationListener(ariaCallback);
        shouldBe("addedNotification", "true");

        // AXLiveRegionsChanged notifications are sent after a delay, so in order to ensure we 
        // get an notification for each operation we should avoid coalescing which can occur
        setTimeout("textChangeOperation()", 0);
        setTimeout("newElementOperation()", 30);
        setTimeout("alternativeChangeOperation()", 60);
        setTimeout("removeElementOperation()", 90);

        // add a short delay to check after all the operations if we have exactly four notificaitons
        setTimeout("finishTest()", 200);
    }

    function textChangeOperation() {
        // this should trigger our live region callback for a text change.
        document.getElementById("liveregion").childNodes[0].innerText = "changed text";
    }

    function newElementOperation() {
        // this should trigger our live region callback for a new element.
        document.getElementById("liveregion").innerHTML += "new text element";
    }

    function alternativeChangeOperation() {
        // this should also trigger our live region change because its a text alternative change. 
        document.getElementById("image").setAttribute('alt', "new image text");

        // Access the accessibility tree here because AX won't post the live region change
        // notification twice when the children have already been marked as dirty.
        liveRegion.childAtIndex(0);
    }

    function removeElementOperation() {
        // this should trigger our live region callback for a removed element.
        document.getElementById("liveregion").removeChild(document.getElementById("innerlive")); 
    }

    function finishTest() {
        // We should get a total of four live region changes.
        liveRegion.removeNotificationListener();
        shouldBeTrue("liveRegionChangeCount == 4");
        finishJSTest();
    }

</script>

<script src="../../../resources/js-test-post.js"></script>
</body>
</html>
